Summarizer/requirements.txt
SQLAlchemy>=2.0.0
slack-bolt==1.18.0
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==21.0.1